requirements.txt
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework_simplejwt==5.4.0
django-oauth-toolkit==3.0.1
drf-spectacular==0.28.0
black==25.1.0
celery[rabbitmq]==5.4.0
pika==1.3.2
django-extensions==4.1